requirements.txt
torch
torchvision
gradio>=4.0
pillow
numpy
